added telegrem_bot change type in models from in to bigint

+++ sql_app/models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Boolean, Column, ForeignKey, Integer, String
+from sqlalchemy import Boolean, Column, ForeignKey, Integer, String, BigInteger
 from sqlalchemy.orm import relationship
 
 from .database import Base
@@ -19,6 +19,6 @@ class User(Base):
 
     id = Column(Integer, primary_key=True, index=True)
     nickname = Column(String(50))
-    chat_id = Column(Integer)
+    chat_id = Column(BigInteger)
 
     message = relationship("Message", backref="users", cascade="all, delete", passive_deletes=True)
